Fix bug in pos_tagging.py: POS-tagger info must be list of dicts
Before: POS-tagger info was stored as a dict of the form {word: POStag} under the key "postagger" in the output json.
Problem: Duplicate words are only stored once (even though their POStags might be different.
Also: This does not preserve the order of the words.
Now: POS-tagger info is stored as list of dicts: [ {word1: POStag1}, {word2: POStag2},...]  under the key "postagger".